% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncvs_api.R
\name{ncvs_api}
\alias{ncvs_api}
\title{Request Data from the National Crime Victimization Survey (NCVS)}
\usage{
ncvs_api(year = 2016, dataset = "personal", population = FALSE, ...)
}
\arguments{
\item{year}{These data archives span all the way back from 1993 to 2016. The default is 2016}

\item{dataset}{The API provides two different datasets, the personal victimization, and household victimization for all years.}

\item{population}{A boolean value specifying whether to return personal victimization population of incidents reported to the NCVS by year.}

\item{...}{Any additional parameters}
}
\value{
A tibble containing information downloaded from API.
}
\description{
Returns data from NCVS API
}
\examples{

# A request for personal victimization survey for the year 2012
dat <- ncvs_api(year = 2012, dataset = "personal", population = FALSE)

# A requestion for household victimization survey for the year 1994
dat <- ncvs_api(year = 1994, dataset = "household", population = FALSE)


}
\seealso{
\url{https://www.bjs.gov/developer/ncvs/developers.cfm#/bjs/ncvs/v2/personal/population/{year}}
}
\keyword{bjs}
\keyword{crime}
\keyword{law}
\keyword{policy}
\keyword{victimization}
